First release
It basically asks you for a search string, which it then searches for on
Wikipedia, and displays the first paragraph of the searched query.
If the search results do not exist, then it just prints out the heading
of the first search suggestion link from Wikipedia.
This is the initial release, and contains many errors, bugs, redundant
code, and what not. Will update this later.